pyproject.toml
[project]
name = "bigquery-agent-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
